Begin coding challenge
... spent about 3.5 hours on the Smarterer coding challenge tonight as
an exercise in learning Python using Flask and SQLAlchemy. Further
improvements to the environment provisioning: got the server and the
database (postgres) to be visible to the host, and began playing around
with stubbing out some of the basic calls that need to be made. We'll
allow the root URL to be a "friendly" instructional view of how to
interact with the service.

Next up? Read up a little more on Python modules and figure out why how
I can separate my logic a little bit better as I've currently got it
laid out: /config for configuration, /models for data models.