Update code, get database and models working
... in general this is a pretty successful session. I spent about 2.5
hours on fixing my module import problems (which, after some trial and
error I realized stemmed from some poor naming on my part and ultimately
decided to move the application run code out of the /app directory).
Following that I was able to make real progress with writing a basic
data importer (database.py) to populate the postgresql database backend
and begin to write the controller logic. There's still some work to be
done to make this fully functional, and unit tests would be nice to have
in place as well.

Learned a bit about list comprehension in the process.

These commit messages are becoming a bit... verbose.